Open/Save local commands should only be contributed in remote
Fixes #75737
