+++ src/vs/workbench/browser/actions/workspaceActions.ts
@@ -15,11 +15,7 @@ import { ICommandService } from 'vs/platform/commands/common/commands';
 import { ADD_ROOT_FOLDER_COMMAND_ID, ADD_ROOT_FOLDER_LABEL, PICK_WORKSPACE_FOLDER_COMMAND_ID } from 'vs/workbench/browser/actions/workspaceCommands';
 import { IFileDialogService } from 'vs/platform/dialogs/common/dialogs';
 import { INotificationService } from 'vs/platform/notification/common/notification';
-import { Schemas } from 'vs/base/common/network';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
-import { ITextFileService, ISaveOptions } from 'vs/workbench/services/textfile/common/textfiles';
-import { toResource } from 'vs/workbench/common/editor';
-import { URI } from 'vs/base/common/uri';
 
 export class OpenFileAction extends Action {
 
@@ -36,47 +32,6 @@ export class OpenFileAction extends Action {
 
 	run(event?: any, data?: ITelemetryData): Promise<any> {
 		return this.dialogService.pickFileAndOpen({ forceNewWindow: false, telemetryExtraData: data });
-	}
-}
-
-export class OpenLocalFileAction extends Action {
-
-	static readonly ID = 'workbench.action.files.openLocalFile';
-	static LABEL = nls.localize('openLocalFile', "Open Local File...");
-
-	constructor(
-		id: string,
-		label: string,
-		@IFileDialogService private readonly dialogService: IFileDialogService
-	) {
-		super(id, label);
-	}
-
-	run(event?: any, data?: ITelemetryData): Promise<any> {
-		return this.dialogService.pickFileAndOpen({ forceNewWindow: false, telemetryExtraData: data, availableFileSystems: [Schemas.file] });
-	}
-}
-
-export class SaveLocalFileAction extends Action {
-
-	static readonly ID = 'workbench.action.files.saveLocalFile';
-	static LABEL = nls.localize('saveLocalFile', "Save Local File...");
-
-	constructor(
-		id: string,
-		label: string,
-		@ITextFileService private readonly textFileService: ITextFileService,
-		@IEditorService private readonly editorService: IEditorService
-	) {
-		super(id, label);
-	}
-
-	async run(event?: any, data?: ITelemetryData): Promise<any> {
-		let resource: URI | undefined = toResource(this.editorService.activeEditor);
-		const options: ISaveOptions = { force: true, availableFileSystems: [Schemas.file] };
-		if (resource) {
-			return this.textFileService.saveAs(resource, undefined, options);
-		}
 	}
 }
 
@@ -98,25 +53,6 @@ export class OpenFolderAction extends Action {
 	}
 }
 
-export class OpenLocalFolderAction extends Action {
-
-	static readonly ID = 'workbench.action.files.openLocalFolder';
-	static LABEL = nls.localize('openLocalFolder', "Open Local Folder...");
-
-	constructor(
-		id: string,
-		label: string,
-		@IFileDialogService private readonly dialogService: IFileDialogService
-	) {
-		super(id, label);
-	}
-
-	run(event?: any, data?: ITelemetryData): Promise<any> {
-		return this.dialogService.pickFolderAndOpen({ forceNewWindow: false, telemetryExtraData: data, availableFileSystems: [Schemas.file] });
-	}
-}
-
-
 export class OpenFileFolderAction extends Action {
 
 	static readonly ID = 'workbench.action.files.openFileFolder';
@@ -132,24 +68,6 @@ export class OpenFileFolderAction extends Action {
 
 	run(event?: any, data?: ITelemetryData): Promise<any> {
 		return this.dialogService.pickFileFolderAndOpen({ forceNewWindow: false, telemetryExtraData: data });
-	}
-}
-
-export class OpenLocalFileFolderAction extends Action {
-
-	static readonly ID = 'workbench.action.files.openLocalFileFolder';
-	static LABEL = nls.localize('openLocalFileFolder', "Open Local...");
-
-	constructor(
-		id: string,
-		label: string,
-		@IFileDialogService private readonly dialogService: IFileDialogService
-	) {
-		super(id, label);
-	}
-
-	run(event?: any, data?: ITelemetryData): Promise<any> {
-		return this.dialogService.pickFileFolderAndOpen({ forceNewWindow: false, telemetryExtraData: data, availableFileSystems: [Schemas.file] });
 	}
 }
